Enforce PR on CI
I know it can be controversial, but I would always protect the main branch, so that changes to it must be made via a PR and any checks must pass before merging. This can be painful for small changes (and can be avoided by administrators), but I think the benefit gained by always having the HEAD of main in a state of green is worth it.